Finally moved Selectors out of ELement.py
