+++ Element.py
@@ -8,14 +8,12 @@ This includes:
 """
 import asyncio
 from contextlib import suppress
-import re
 from dataclasses import dataclass
-from functools import cache, cached_property, reduce
+from functools import cache
 from itertools import chain
 
 # fmt: off
-from pprint import pprint
-from typing import (TYPE_CHECKING, Any, Callable, Iterable, Optional, Protocol, Sequence,
+from typing import (TYPE_CHECKING, Callable, Iterable, Optional, Sequence,
                     Type, Union)
 
 import pygame as pg
@@ -31,27 +29,12 @@ from own_types import (Auto, AutoLP4Tuple, AutoType, BugError, Color, Coordinate
 from Style import (SourceSheet, bc_getter, br_getter, bs_getter, bw_keys,
                    get_style, inset_getter, pack_longhands, parse_file,
                    parse_sheet, prio_keys, calculator)
-from util import create_task, draw_text, get_groups, get_tag, group_by_bool, log_error
+from util import create_task, draw_text, get_tag, group_by_bool, log_error
 
 # fmt: on
 """ More useful links for further development
 https://developer.mozilla.org/en-US/docs/Web/CSS/image
 """
-
-########################## Specificity and Rules #############################
-Spec = tuple[int, int, int]
-
-
-def add_specs(t1: Spec, t2: Spec) -> Spec:
-    """
-    Cumulate two Specificities
-    """
-    id1, cls1, tag1 = t1
-    id2, cls2, tag2 = t2
-    return (id1 + id2, cls1 + cls2, tag1 + tag2)
-
-
-sum_specs: Callable[[Iterable[Spec]], Spec] = lambda specs: reduce(add_specs, specs)
 
 ################################# Globals ###################################
 
@@ -137,9 +120,7 @@ class Element:
         if TYPE_CHECKING:
             assert isinstance(parent, Element)
         self.parent = parent
-        # parse element style and update default
         self.istyle = Style.parse_inline_style(attrs.get("style", ""))
-        self.estyle = {}
 
     def is_block(self) -> bool:
         """
@@ -175,11 +156,6 @@ class Element:
         return Style.remove_importantd(Style.join_styles(self.istyle, self.estyle))
 
     ####################################  Main functions ######################################################
-    @cache
-    def matches(self, selector: "Selector"):
-        """Returns whether the given Selector matches the Element, cached"""
-        return selector(self)
-
     def collide(self, pos: Coordinate) -> Iterable["Element"]:
         """
         The idea of this function is to get which elements were hit for focus, hover, etc.
@@ -389,10 +365,6 @@ class Element:
                         item.element.cstyle["color"],
                         topleft=text_pos,
                     )
-                    # word_surf = self.font.render(
-                    #     item.text, True, item.element.cstyle["color"]
-                    # )
-                    # surf.blit(word_surf, pos)
                 ypos += line.height
         else:
             raise BugError(f"Wrong layout_type ({self.layout_type})")
@@ -825,15 +797,12 @@ def apply_style():
     """
     Apply the global SourceSheet to all elements
     """
-    g["global_sheet"] = SourceSheet.join(g["css_sheets"])
+    joined_sheet = SourceSheet.join(g["css_sheets"])
     g["css_dirty"] = False
     g["css_sheet_len"] = len(g["css_sheets"])
     # sort all rules by the selectors specificities
     rules = sorted(
-        (
-            (parse_selector(str_sel), style)
-            for str_sel, style in g["global_sheet"].all_rules
-        ),
+        joined_sheet.all_rules,
         key=lambda rule: rule[0].spec,
     )
     root: Element = g["root"]
@@ -847,265 +816,3 @@ def apply_style():
                 key=lambda t: Style.is_imp(t[1]),
             )
         )
-
-
-################################# Selectors #######################################
-# https://www.w3.org/TR/selectors-3/
-"""
-Handles everything revolving Selectors. 
-Includes parsing (text->Selector)
-"""
-
-Selector = Union["SingleSelector", "CompositeSelector"]
-
-
-class SingleSelector(Protocol):
-    spec: Spec
-
-    def __call__(self: Any, elem: Element) -> bool:
-        ...
-
-    def __hash__(self) -> int:
-        return super().__hash__()
-
-
-@dataclass(frozen=True, slots=True)
-class TagSelector:
-    tag: str
-    spec = 0, 0, 1
-    __call__ = lambda self, elem: elem.tag == self.tag
-    __str__ = lambda self: self.tag  # type: ignore[attr-defined]
-
-
-@dataclass(frozen=True, slots=True)
-class IdSelector:
-    id: str
-    spec = 1, 0, 0
-    __call__ = lambda self, elem: elem.attrs.get("id") == self.id
-    __str__ = lambda self: "#" + self.id  # type: ignore[attr-defined]
-
-
-@dataclass(frozen=True, slots=True)
-class ClassSelector:
-    cls: str
-    spec = 0, 1, 0
-    __call__ = lambda self, elem: elem.attrs.get("class") == self.cls
-    __str__ = lambda self: "." + self.cls  # type: ignore[attr-defined]
-
-
-@dataclass(frozen=True, slots=True)
-class HasAttrSelector:
-    name: str
-    spec = 0, 1, 0
-    __call__ = lambda self, elem: self.name in elem.attrs
-    __str__ = lambda self: f"[{self.name}]"  # type: ignore[attr-defined]
-
-
-# the validator takes the soll value and the is value
-def make_attr_selector(sign: str, validator):
-    sign = re.escape(sign)
-    regex = re.compile(
-        rf'\[{s}(\w+){s}{sign}={s}"(\w+)"{s}\]|\[{s}(\w+){s}{sign}={s}(\w+){s}\]'
-    )
-
-    @dataclass(frozen=True, slots=True)
-    class AttributeSelector:
-        name: str
-        value: str
-        spec = 0, 1, 0
-
-        def __call__(self, elem: Element):
-            return (value := elem.attrs.get(self.name)) is not None and validator(
-                self.value, value
-            )
-
-        def __str__(self):
-            return f'[{self.name}{sign}="{self.value}"]'
-
-    return regex, AttributeSelector
-
-
-@dataclass(frozen=True, slots=True)
-class AnySelector:
-    spec = 0, 0, 0
-
-    def __call__(self, elem: Element):
-        return True
-
-    def __str__(self):
-        return "*"
-
-
-@dataclass(frozen=True, slots=True)
-class NeverSelector:
-    spec: Spec
-    s: str
-
-    def __call__(self, elem: Element):
-        return False
-
-    def __str__(self):
-        return self.s
-
-
-################################## Composite Selectors ###############################
-class CompositeSelector:
-    selectors: tuple[Selector, ...]
-    spec: cached_property[Spec]
-
-    def __call__(self: Any, elem: Element) -> bool:
-        ...
-
-    def __hash__(self) -> int:
-        return super().__hash__()
-
-
-joined_specs = lambda self: sum_specs(f.spec for f in self.selectors)
-
-
-@dataclass(frozen=True, slots=True)
-class AndSelector(CompositeSelector):
-    selectors: tuple[Selector, ...]
-    spec = cached_property(joined_specs)
-    __call__ = lambda self, elem: all(elem.matches(sel) for sel in self.selectors)
-    __str__ = lambda self: "".join(str(s) for s in self.selectors)  # type: ignore[attr-defined]
-
-
-@dataclass(frozen=True, slots=True)
-class OrSelector(CompositeSelector):
-    selectors: tuple[Selector, ...]
-    spec = cached_property(joined_specs)
-    __call__ = lambda self, elem: any(elem.matches(sel) for sel in self.selectors)
-    __str__ = lambda self: ", ".join(str(s) for s in self.selectors)  # type: ignore[attr-defined]
-
-
-@dataclass(frozen=True, slots=True)
-class DirectChildSelector(CompositeSelector):
-    selectors: tuple[Selector, ...]
-    spec = cached_property(joined_specs)
-
-    def __call__(self, elem: Element):
-        chain = [elem, *elem.iter_anc()]
-        if len(chain) != len(self.selectors):
-            return False
-        return all(parent.matches(sel) for parent, sel in zip(chain, self.selectors))
-
-    __str__ = lambda self: " > ".join(str(s) for s in self.selectors)  # type: ignore[attr-defined]
-
-
-@dataclass(frozen=True, slots=True)
-class ChildSelector(CompositeSelector):
-    selectors: tuple[Selector, Selector]
-    spec = cached_property(joined_specs)
-
-    def __call__(self, elem: Element):
-        own_sel, p_sel = self.selectors
-        if not own_sel(elem):
-            return False
-        return any(p.matches(p_sel) for p in elem.iter_anc())
-
-    __str__ = lambda self: " ".join(str(s) for s in self.selectors)  # type: ignore[attr-defined]
-
-
-########################################## Parser #######################################################
-s = r"\s*"
-sngl_p = re.compile(
-    r"((?:\*|(?:#\w+)|(?:\.\w+)|(?:\[\s*\w+\s*\])|(?:\[\s*\w+\s*[~|^$*]?=\s*\w+\s*\])|(?:\w+)))$"
-)
-rel_p = re.compile(r"\s*([>+~ ])\s*$")  # pretty simple
-
-attr_sel_data = [
-    ("", lambda soll, _is: soll == _is),
-    ("~", lambda soll, _is: soll in _is.split()),
-    ("|", lambda soll, _is: soll == _is or _is.startswith(soll + "-")),
-    ("^", lambda soll, _is: _is.startswith(soll)),
-    ("$", lambda soll, _is: _is.endswith(soll)),
-    ("*", lambda soll, _is: soll in _is),
-]
-attr_patterns: list[tuple[re.Pattern, Type[Selector]]] = [
-    (re.compile(r"\[(\w+)\]"), HasAttrSelector),
-    *(make_attr_selector(sign, validator) for sign, validator in attr_sel_data),
-]
-
-
-def matches(s: str, pattern: re.Pattern):
-    """
-    Takes a string and a pattern and returns the reststring and the capture
-    """
-    if not ((match := pattern.search(s)) and (length := len(match.group()))):
-        return None
-    if not (groups := [g for g in match.groups() if g]):
-        return None
-    return (s[:-length], groups[0])
-
-
-class InvalidSelector(ValueError):
-    pass
-
-
-@cache
-def parse_selector(s: str) -> Selector:
-    s = start = s.strip()
-    if not s:
-        raise InvalidSelector("Empty selector")
-    if "," in s:
-        return OrSelector(tuple(parse_selector(x) for x in s.split(",")))
-    else:
-        singles: list[str] = []
-        # do-while-loop
-        while True:
-            if match := matches(s, sngl_p):
-                s, subsel = match
-                singles.insert(0, subsel)
-                if not s:  # recursive anker
-                    return proc_singles(singles)
-            elif match := matches(s, rel_p):
-                s, rel = match
-                if not s or not singles:
-                    raise InvalidSelector(
-                        f"Relative selectors are not padded by single selectors: ({rel},{start})"
-                    )
-                sngl_selector = proc_singles(singles)
-                match rel:
-                    case " ":
-                        return ChildSelector((parse_selector(s), sngl_selector))
-                    case ">":
-                        return DirectChildSelector((parse_selector(s), sngl_selector))
-                    case "+":
-                        raise NotImplementedError("+ is not implemented yet")
-                    case "~":
-                        raise NotImplementedError("~ is not implemented yet")
-                    case _:
-                        raise BugError(f"Invalid relative selector ({rel})")
-            else:
-                raise InvalidSelector(f"Couldn't match '{s}'")
-
-
-def proc_singles(groups: list[str]) -> Selector:
-    """
-    Merge several selectors into one
-    """
-    if len(groups) == 1:
-        return proc_single(groups[0])
-    return AndSelector(tuple(proc_single(s) for s in groups))
-
-
-def proc_single(s: str) -> Selector:
-    """
-    Create a single selector from a string
-    """
-    if s == "*":
-        return AnySelector()
-    elif s[0] == "#":
-        return IdSelector(s[1:])
-    elif s[0] == ".":
-        return ClassSelector(s[1:])
-    elif s[0] == "[":
-        for pattern, selector in attr_patterns:
-            if (groups := get_groups(s, pattern)) is not None:
-                return selector(*groups)
-        raise BugError(f"Invalid attribute selector: {s}")
-    elif s[0] == ":":  # pseudoclass
-        raise RuntimeError("Pseudoclasses are not supported yet")
-    else:
-        return TagSelector(s)
