+++ own_types.py
@@ -9,7 +9,7 @@ from functools import reduce
 from operator import or_
 
 # fmt: off
-from typing import (Generator, Generic, Hashable, Iterable, Literal,
+from typing import (Any, Generator, Generic, Hashable, Iterable, Literal,
                     Mapping, Protocol, TypeVar, Union)
 # fmt: on
 from weakref import WeakValueDictionary
@@ -57,6 +57,18 @@ CO_T = TypeVar("CO_T", covariant=True)
 class Drawable(Protocol):
     def draw(self, surf: Surface, pos: Coordinate):
         pass
+
+
+class Element_P(Protocol):
+    """
+    A Protocol that just represents an Element
+    """
+
+    tag: str
+    attrs: dict[str, Any]
+
+    def iter_anc(self) -> Iterable["Element_P"]:
+        ...
 
 
 class Enum(_Enum):
